EmsMemoryMapper class

+++ src/Spice86.Core/Emulator/ReverseEngineer/MemoryBasedDataStructureWithBaseAddressProvider.cs
@@ -59,12 +59,12 @@ public abstract class MemoryBasedDataStructureWithBaseAddressProvider : MemoryBa
         SetUint8(BaseAddress, offset, value);
     }
 
-    public void SetZeroTerminatedString(int start, string value, int maxLength) {
+    public void SetZeroTerminatedString(uint start, string value, int maxLength) {
         if (value.Length + 1 > maxLength) {
             throw new UnrecoverableException($"String {value} is more than {maxLength} cannot write it at offset {start}");
         }
 
-        uint physicalStart = (uint)(BaseAddress + start);
+        uint physicalStart = (BaseAddress + start);
         int i = 0;
         for (; i < value.Length; i++) {
             char character = value[i];
